# Backend/openapi.yaml
openapi: 3.0.3
info:
  title: WealthTrack Microservice Ecosystem API
  version: 1.0.0
  description: |
    Code-derived API contract for every custom controller outside `MAIN`.

    **Authentication:** no Spring Security configuration, authentication filter, or
    authorization policy exists in the reviewed services. Every operation therefore
    declares `security: []`. The reusable 401 and 403 responses are future-security
    templates only and are not currently emitted by application authentication.

    **Headers:** JSON operations accept `Content-Type: application/json` and normally
    return `application/json`; clients may send `Accept: application/json`. Delete and
    internal command operations return no body. There are no application-specific
    request headers in the source code.

    **Gateway use:** each business path has a direct-service server and an API Gateway
    server. The gateway removes `/user`, `/bank`, `/portfolio`, `/trading`, or
    `/product` before forwarding. Internal endpoints are not access-controlled and are
    also reachable through the gateway as currently configured.

    **Uncertain/framework contracts:** Eureka exposes its library-managed dashboard and
    discovery protocol under `/eureka/**`, but the repository defines no custom Eureka
    controller or DTO contract. Those framework endpoints are described in
    `MICROSERVICE_API_MAP.md` and deliberately not invented here.
servers:
  - url: http://localhost:8081
...
security: []
paths:
...
  /api/portfolios:
    servers:
      - url: http://localhost:8084
        description: Portfolio Service direct
      - url: http://localhost:8081/portfolio
        description: Portfolio Service through API Gateway
    post:
      tags: [Portfolio]
      operationId: createPortfolioAccount
      summary: Open a portfolio account
      description: Validates the referenced ACTIVE user through User Service, enforces one portfolio per user, rejects future opened dates, and defaults openedDate to today.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePortfolioAccountRequest' }
            example: { userId: 10, openedDate: '2026-07-30' }
      responses:
        '201':
          description: Portfolio account opened
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortfolioAccountResponse' }
              example:
                portfolioAccountId: 1
                userId: 10
                accountStatus: ACTIVE
                openedDate: '2026-07-30'
                closedDate: null
                createdAt: '2026-07-30T10:50:00'
                updatedAt: '2026-07-30T10:50:00'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    get:
      tags: [Portfolio]
      operationId: listPortfolioAccounts
      summary: List portfolio accounts
      description: Returns all portfolio accounts without filtering or pagination.
      security: []
      responses:
        '200':
          description: Portfolio account list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PortfolioAccountResponse' }
              example:
                - portfolioAccountId: 1
                  userId: 10
                  accountStatus: ACTIVE
                  openedDate: '2026-07-30'
                  closedDate: null
                  createdAt: '2026-07-30T10:50:00'
                  updatedAt: '2026-07-30T10:50:00'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/portfolios/{portfolioAccountId}:
    servers:
      - url: http://localhost:8084
        description: Portfolio Service direct
      - url: http://localhost:8081/portfolio
        description: Portfolio Service through API Gateway
    get:
      tags: [Portfolio]
      operationId: getPortfolioAccount
      summary: Get a portfolio account
      description: Returns account status and lifecycle dates.
      security: []
      parameters:
        - $ref: '#/components/parameters/PortfolioAccountId'
      responses:
        '200':
          description: Portfolio account found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortfolioAccountResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/portfolios/by-user/{userId}:
    servers:
      - url: http://localhost:8084
        description: Portfolio Service direct
      - url: http://localhost:8081/portfolio
        description: Portfolio Service through API Gateway
    get:
      tags: [Portfolio]
      operationId: getPortfolioAccountByUser
      summary: Get a user's portfolio account
      description: Finds the single portfolio for userId. Trading consumes this endpoint for investment overview aggregation.
      security: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Portfolio account found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortfolioAccountResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/portfolios/{portfolioAccountId}/status:
    servers:
      - url: http://localhost:8084
        description: Portfolio Service direct
      - url: http://localhost:8081/portfolio
        description: Portfolio Service through API Gateway
    patch:
      tags: [Portfolio]
      operationId: updatePortfolioStatus
      summary: Change portfolio account status
      description: Sets ACTIVE, SUSPENDED, or CLOSED. Closing is rejected while any non-closed holding has positive quantity; closedDate is maintained automatically.
      security: []
      parameters:
        - $ref: '#/components/parameters/PortfolioAccountId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdatePortfolioStatusRequest' }
            example: { accountStatus: SUSPENDED }
      responses:
        '200':
          description: Updated portfolio account
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortfolioAccountResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/portfolios/{portfolioAccountId}/holdings:
    servers:
      - url: http://localhost:8084
        description: Portfolio Service direct
      - url: http://localhost:8081/portfolio
        description: Portfolio Service through API Gateway
    post:
      tags: [Portfolio]
      operationId: addPortfolioHolding
      summary: Add to a portfolio holding
      description: |
        Requires an ACTIVE portfolio and an ACTIVE product validated through Product
        Service. Creates a holding or increases the existing product holding, calculates
        weighted average cost, and values it using that average cost.
      security: []
      parameters:
        - $ref: '#/components/parameters/PortfolioAccountId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateHoldingRequest' }
            example: { productId: 1, quantity: 100.0000, averageCost: 1400.0000 }
      responses:
        '201':
          description: Holding created or increased
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortfolioHoldingResponse' }
              example:
                holdingId: 1
                portfolioAccountId: 1
                productId: 1
                quantity: 100.0000
                averageCost: 1400.0000
                marketValue: 140000.0000
                unrealizedGainLoss: 0.0000
                holdingStatus: ACTIVE
                lastValuedAt: '2026-07-30T10:55:00'
                createdAt: '2026-07-30T10:55:00'
                updatedAt: '2026-07-30T10:55:00'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    get:
      tags: [Portfolio]
      operationId: listPortfolioHoldings
      summary: List portfolio holdings
      description: Validates that the portfolio exists and optionally filters holdings by status. Trading consumes the unfiltered form for investment overview.
      security: []
      parameters:
        - $ref: '#/components/parameters/PortfolioAccountId'
        - name: status
          in: query
          required: false
          schema: { $ref: '#/components/schemas/HoldingStatus' }
          example: ACTIVE
      responses:
        '200':
          description: Portfolio holdings
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PortfolioHoldingResponse' }
              example:
                - holdingId: 1
                  portfolioAccountId: 1
                  productId: 1
                  quantity: 100.0000
                  averageCost: 1400.0000
                  marketValue: 140000.0000
                  unrealizedGainLoss: 0.0000
                  holdingStatus: ACTIVE
                  lastValuedAt: '2026-07-30T10:55:00'
                  createdAt: '2026-07-30T10:55:00'
                  updatedAt: '2026-07-30T10:55:00'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/portfolios/holdings/{holdingId}:
    servers:
      - url: http://localhost:8084
        description: Portfolio Service direct
      - url: http://localhost:8081/portfolio
        description: Portfolio Service through API Gateway
    parameters:
      - $ref: '#/components/parameters/HoldingId'
    get:
      tags: [Portfolio]
      operationId: getPortfolioHolding
      summary: Get a holding
      description: Returns quantity, cost, stored market value, calculated unrealized gain/loss, status, and timestamps.
      security: []
      responses:
        '200':
          description: Holding found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortfolioHoldingResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    patch:
      tags: [Portfolio]
      operationId: updatePortfolioHolding
      summary: Partially update a holding
      description: Updates non-null quantity, average cost, or holding status on an ACTIVE portfolio. Quantity zero automatically closes and zero-values the holding.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateHoldingRequest' }
            example: { quantity: 105.0000, averageCost: 1402.3810, holdingStatus: ACTIVE }
      responses:
        '200':
          description: Updated holding
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortfolioHoldingResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    delete:
      tags: [Portfolio]
      operationId: deletePortfolioHolding
      summary: Delete an empty or closed holding
      description: Permanently deletes a holding only if it has no remaining quantity or is already CLOSED.
      security: []
      responses:
        '204': { description: Holding deleted; no response body }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/portfolios/internal/trades/validate:
    servers:
      - url: http://localhost:8084
        description: Portfolio Service direct
      - url: http://localhost:8081/portfolio
        description: Portfolio Service through API Gateway
    post:
      tags: [Portfolio - Internal]
      operationId: validatePortfolioTrade
      summary: Pre-validate a trade command
      description: |
        Internal synchronous REST contract consumed by Trading before money moves.
        Confirms active portfolio, matching portfolio/holding/product IDs, BUY or SELL
        type, and sufficient SELL quantity. Despite its internal role, no authentication
        or gateway exclusion protects it.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApplyTradeRequest' }
            example:
              portfolioAccountId: 1
              holdingId: 1
              productId: 1
              transactionType: BUY
              quantity: 1.0000
              unitPrice: 1450.0000
              transactionId: 8
      responses:
        '204': { description: Trade is valid; no response body }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/portfolios/internal/trades:
    servers:
      - url: http://localhost:8084
        description: Portfolio Service direct
      - url: http://localhost:8081/portfolio
        description: Portfolio Service through API Gateway
    post:
      tags: [Portfolio - Internal]
      operationId: applyCompletedPortfolioTrade
      summary: Apply an already-funded trade
      description: |
        Internal command used by Trading after the corresponding bank step. Re-runs
        validation, then increases quantity and weighted average cost for BUY or
        decreases quantity for SELL. The supplied unitPrice sets marketValue. The
        transactionId is validated as positive but is not persisted or deduplicated,
        so the endpoint is not idempotent.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApplyTradeRequest' }
            example:
              portfolioAccountId: 1
              holdingId: 1
              productId: 1
              transactionType: BUY
              quantity: 1.0000
              unitPrice: 1450.0000
              transactionId: 8
      responses:
        '204': { description: Trade applied; no response body }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/portfolios/{portfolioAccountId}/summary:
    servers:
      - url: http://localhost:8084
        description: Portfolio Service direct
      - url: http://localhost:8081/portfolio
        description: Portfolio Service through API Gateway
    get:
      tags: [Portfolio]
      operationId: getPortfolioSummary
      summary: Get a portfolio valuation summary
      description: Returns the account, all holdings, total cost, summed stored market value, and unrealized gain/loss. It does not fetch fresh Product prices.
      security: []
      parameters:
        - $ref: '#/components/parameters/PortfolioAccountId'
      responses:
        '200':
          description: Portfolio summary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortfolioSummaryResponse' }
              example:
                portfolioAccount:
                  portfolioAccountId: 1
                  userId: 10
                  accountStatus: ACTIVE
                  openedDate: '2026-07-30'
                  closedDate: null
                  createdAt: '2026-07-30T10:50:00'
                  updatedAt: '2026-07-30T10:50:00'
                holdings:
                  - holdingId: 1
                    portfolioAccountId: 1
                    productId: 1
                    quantity: 101.0000
                    averageCost: 1400.4950
                    marketValue: 146450.0000
                    unrealizedGainLoss: 4999.9950
                    holdingStatus: ACTIVE
                    lastValuedAt: '2026-07-30T11:00:00'
                    createdAt: '2026-07-30T10:55:00'
                    updatedAt: '2026-07-30T11:00:00'
                totalCost: 141449.9950
                marketValue: 146450.0000
                unrealizedGainLoss: 5000.0050
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
components:
  parameters:
...
    UserId:
      name: userId
      in: path
      required: true
      description: User identifier
      schema: { type: integer, format: int64, minimum: 1 }
      example: 10
...
    PortfolioAccountId:
      name: portfolioAccountId
      in: path
      required: true
      description: Portfolio account identifier
      schema: { type: integer, format: int64, minimum: 1 }
      example: 1
    HoldingId:
      name: holdingId
      in: path
      required: true
      description: Portfolio holding identifier
      schema: { type: integer, format: int64, minimum: 1 }
      example: 1
  responses:
    BadRequest:
      description: Request validation, enum parsing, date-range, or business-rule failure
      content:
        application/json:
          schema: { $ref: '#/components/schemas/StandardApiError' }
          examples:
            validation:
              summary: Validation error
              value:
                timestamp: '2026-07-30T11:15:00'
                status: 400
                error: Bad Request
                message: Request validation failed
                path: /api/bank-accounts
                fieldErrors:
                  accountNumber: must not be blank
            businessRule:
              summary: Business-rule error
              value:
                timestamp: '2026-07-30T11:15:00'
                status: 400
                message: Portfolio account 1 is not active
    Unauthorized:
      description: |
        Reusable 401 template. No reviewed service currently authenticates requests, so
        this response is not emitted by application security in the current code.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/StandardApiError' }
          example:
            timestamp: '2026-07-30T11:15:00'
            status: 401
            error: Unauthorized
            message: Authentication is required
            path: /api/example
    Forbidden:
      description: |
        Reusable 403 template. No reviewed service currently enforces roles or
        permissions, so this response is not emitted by application authorization.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/StandardApiError' }
          example:
            timestamp: '2026-07-30T11:15:00'
            status: 403
            error: Forbidden
            message: The authenticated principal cannot perform this operation
            path: /api/example
    NotFound:
      description: Requested resource does not exist
      content:
        application/json:
          schema: { $ref: '#/components/schemas/StandardApiError' }
          example:
            timestamp: '2026-07-30T11:15:00'
            status: 404
            error: Not Found
            message: Resource not found
            path: /api/example/999
...
    InternalServerError:
      description: |
        Unexpected server or dependency failure. Bank and Portfolio define this envelope
        explicitly. User, Product, Trading, and Gateway may fall through to Spring's
        default error handling, so their exact 500 body is not fully determined.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/StandardApiError' }
          example:
            timestamp: '2026-07-30T11:15:00'
            status: 500
            error: Internal Server Error
            message: Internal server error
            path: /api/example
...
  schemas:
    StandardApiError:
      type: object
      description: |
        Superset reusable error envelope covering the three code-defined formats. Only
        `timestamp`, `status`, and `message` are common to all custom handlers. Optional
        fields appear in Bank/Portfolio handlers. Spring default errors can contain
        additional fields.
      required: [timestamp, status, message]
      properties:
        timestamp:
          type: string
          format: date-time
          example: '2026-07-30T11:15:00'
        status:
          type: integer
          format: int32
          example: 400
        error:
          type: string
          nullable: true
          example: Bad Request
        message:
          type: string
          example: Request validation failed
        path:
          type: string
          nullable: true
          example: /api/bank-accounts
        fieldErrors:
          type: object
          nullable: true
          additionalProperties: { type: string }
          example: { accountNumber: must not be blank }
        validationErrors:
          type: object
          nullable: true
          additionalProperties: { type: string }
          example: { quantity: must be greater than or equal to 0.0001 }
...
    AccountStatus:
      type: string
      enum: [ACTIVE, SUSPENDED, CLOSED]
    HoldingStatus:
      type: string
      enum: [ACTIVE, MATURED, CLOSED]
    CreatePortfolioAccountRequest:
      type: object
      required: [userId]
      properties:
        userId: { type: integer, format: int64, example: 10 }
        openedDate: { type: string, format: date, nullable: true, example: '2026-07-30' }
    UpdatePortfolioStatusRequest:
      type: object
      required: [accountStatus]
      properties:
        accountStatus: { $ref: '#/components/schemas/AccountStatus' }
    PortfolioAccountResponse:
      type: object
      required: [portfolioAccountId, userId, accountStatus, openedDate, createdAt, updatedAt]
      properties:
        portfolioAccountId: { type: integer, format: int64, example: 1 }
        userId: { type: integer, format: int64, example: 10 }
        accountStatus: { $ref: '#/components/schemas/AccountStatus' }
        openedDate: { type: string, format: date, example: '2026-07-30' }
        closedDate: { type: string, format: date, nullable: true, example: null }
        createdAt: { type: string, format: date-time, example: '2026-07-30T10:50:00' }
        updatedAt: { type: string, format: date-time, example: '2026-07-30T10:50:00' }
    CreateHoldingRequest:
      type: object
      required: [productId, quantity, averageCost]
      properties:
        productId:
          type: integer
          format: int64
          description: "@NotNull is present in source; positivity is not validated by this DTO."
          example: 1
        quantity: { type: number, minimum: 0.0001, example: 100.0000 }
        averageCost: { type: number, minimum: 0, example: 1400.0000 }
    UpdateHoldingRequest:
      type: object
      properties:
        quantity: { type: number, minimum: 0, nullable: true, example: 105.0000 }
        averageCost: { type: number, minimum: 0, nullable: true, example: 1402.3810 }
        holdingStatus:
          allOf: [{ $ref: '#/components/schemas/HoldingStatus' }]
          nullable: true
    PortfolioHoldingResponse:
      type: object
      required: [holdingId, portfolioAccountId, productId, quantity, averageCost, marketValue, unrealizedGainLoss, holdingStatus, lastValuedAt, createdAt, updatedAt]
      properties:
        holdingId: { type: integer, format: int64, example: 1 }
        portfolioAccountId: { type: integer, format: int64, example: 1 }
        productId: { type: integer, format: int64, example: 1 }
        quantity: { type: number, example: 101.0000 }
        averageCost: { type: number, example: 1400.4950 }
        marketValue: { type: number, example: 146450.0000 }
        unrealizedGainLoss: { type: number, example: 5000.0050 }
        holdingStatus: { $ref: '#/components/schemas/HoldingStatus' }
        lastValuedAt: { type: string, format: date-time, example: '2026-07-30T11:00:00' }
        createdAt: { type: string, format: date-time, example: '2026-07-30T10:55:00' }
        updatedAt: { type: string, format: date-time, example: '2026-07-30T11:00:00' }
    ApplyTradeRequest:
      type: object
      description: Portfolio DTO and Trading HoldingUpdateRequest have the same field names and compatible JSON structure.
      required: [portfolioAccountId, holdingId, productId, transactionType, quantity, unitPrice, transactionId]
      properties:
        portfolioAccountId: { type: integer, format: int64, minimum: 1, example: 1 }
        holdingId: { type: integer, format: int64, minimum: 1, example: 1 }
        productId: { type: integer, format: int64, minimum: 1, example: 1 }
        transactionType: { $ref: '#/components/schemas/TransactionType' }
        quantity: { type: number, minimum: 0.0001, example: 1.0000 }
        unitPrice: { type: number, minimum: 0.0001, example: 1450.0000 }
        transactionId: { type: integer, format: int64, minimum: 1, example: 8 }
    PortfolioSummaryResponse:
      type: object
      required: [portfolioAccount, holdings, totalCost, marketValue, unrealizedGainLoss]
      properties:
        portfolioAccount: { $ref: '#/components/schemas/PortfolioAccountResponse' }
        holdings:
          type: array
          items: { $ref: '#/components/schemas/PortfolioHoldingResponse' }
        totalCost: { type: number, example: 141449.9950 }
        marketValue: { type: number, example: 146450.0000 }
        unrealizedGainLoss: { type: number, example: 5000.0050 }
    TransactionType:
      type: string
      enum: [BUY, SELL]
...